+++ src/app/components/conversation-detail/message-text-area/message-text-area.component.ts
@@ -53,7 +53,7 @@ export class MessageTextAreaComponent implements OnInit, AfterViewInit, OnChange
   @Input() leadInfo: {lead_id: string, hasEmail: boolean, email: string, projectId: string, presence: {}};
   @Input() fileUploadAccept: string;
   @Input() emailSection: boolean;
-  @Input() templatesSection: boolean;
+  @Input() whatsappTemplatesSection: boolean;
   @Input() isOpenInfoConversation: boolean;
   @Input() translationMap: Map<string, string>;
   @Input() dropEvent: any;
@@ -409,7 +409,7 @@ export class MessageTextAreaComponent implements OnInit, AfterViewInit, OnChange
         if (form && form.text && form.text.trim() !== '') {
           const text = '**' + form.subject + '**\r\n' + form.text
           const attributes = {
-            channel: TYPE_MSG_EMAIL
+            offline_channel: TYPE_MSG_EMAIL
           }
           this.eventSendMessage.emit({ msg: text, type: TYPE_MSG_TEXT, metadata: null, attributes: attributes });
         }