+++ src/chat21-core/utils/constants.ts
@@ -39,6 +39,12 @@ export const BCK_COLOR_CONVERSATION_UNSELECTED = '#FFFFFF ';
 
 export const TYPE_SUPPORT_GROUP = 'support-group';
 export const CHANNEL_TYPE_GROUP = 'group';
+export const CHANNEL_TYPE_CHAT21 = 'chat21';
+export const CHANNEL_TYPE_EMAIL = 'email';
+export const CHANNEL_TYPE_FORM = 'form';
+export const CHANNEL_TYPE_MESSANGER = 'messanger';
+export const CHANNEL_TYPE_WHATSAPP = 'whatsapp';
+export const CHANNEL_TYPE_TELEGRAM = 'telegram';
 
 // TYPES MESSAGES
 export const TYPE_MSG_TEXT = 'text';