+++ src/environments/environment.prod.ts
@@ -37,5 +37,6 @@ export const environment = {
     supportMode: false,
     writeToButton: false,
     archivedButton: false,
-    emailSection: true
+    emailSection: true,
+    whatsappTemplatesSection: true
 };